ヘッダーを別コンポーネントとして分離
